+++ frontend/src/actions/index.js
@@ -1,9 +1,4 @@
-import {
-  SIGN_IN,
-  SIGN_OUT,
-  FETCH_USER_DETAILS,
-  UPDATE_USER_DETAILS,
-} from "./types";
+import { SIGN_IN, SIGN_OUT } from "./types";
 import api from "../api/api";
 
 const signIn = (formValues) => async (dispatch) => {
@@ -13,7 +8,7 @@ const signIn = (formValues) => async (dispatch) => {
 
 const signOut = () => async (dispatch, getState) => {
   const { userData } = getState().auth;
-  const response = await api.post(
+  await api.post(
     "/auth/logout/",
     { refresh: userData.refresh },
     {
@@ -25,29 +20,4 @@ const signOut = () => async (dispatch, getState) => {
   dispatch({ type: SIGN_OUT });
 };
 
-const getUserDetails = () => async (dispatch, getState) => {
-  const { userData } = getState().auth;
-  const response = await api.get(`/profile/${userData.user_id}/`, {
-    headers: {
-      Authorization: `Bearer ${userData.access}`,
-    },
-  });
-  dispatch({ type: FETCH_USER_DETAILS, payload: response.data });
-};
-
-const updateUserDetails = (formValues) => async (dispatch, getState) => {
-  const { userData } = getState().auth;
-  const response = await api.patch(
-    `/profile/${userData.user_id}/`,
-    formValues,
-    {
-      headers: {
-        "Content-Type": "multipart/form-data",
-        Authorization: `Bearer ${userData.access}`,
-      },
-    }
-  );
-  dispatch({ type: UPDATE_USER_DETAILS, payload: response.data });
-};
-
-export { signIn, signOut, getUserDetails, updateUserDetails };
+export { signIn, signOut };
