+++ frontend/src/components/SignUpForm.js
@@ -24,15 +24,11 @@ const SignUpForm = () => {
   }, [auth.isLoggedIn, history]);
 
   const signUp = (formValues) => {
-    api
-      .post("/signup/", formValues)
+    api.post("/signup/", formValues)
       .then((res) => {
         history.push("/signin");
-        console.log(res.data);
       })
-      .catch((err) => {
-        console.log(Object.values(err.response.data));
-      });
+      .catch((err) => {});
   };
   return (
     <div>
