+++ frontend/src/pages/HomePage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "../component/Navbar";
+import Navbar from "../components/Navbar";
 const HomePage = () => {
   return (
     <div>
