+++ frontend/src/reducers/profileReducer.js
@@ -1,14 +1,11 @@
 import { FETCH_USER_DETAILS, UPDATE_USER_DETAILS } from "../actions/types";
 
-const INITIAL_STATE = {
-  profile: {},
-};
-const profileReducer = (state = INITIAL_STATE, action) => {
+const profileReducer = (state = {}, action) => {
   switch (action.type) {
     case FETCH_USER_DETAILS:
-      return { ...state, profile: action.payload };
+      return { ...state, ...action.payload };
     case UPDATE_USER_DETAILS:
-      return { ...state, profile: action.payload };
+      return { ...state, ...action.payload };
     default:
       return state;
   }
